TaskStore.cpp: include .cup files by default.
"More" button will add .IGC files.  Theory is that users may have lots of IGC files
with tasks embedded that the user does not normally want to see.  But, that
users may export .CUP files with tasks they make in CU that are what they
like to use in XCSoar.