TaskRenderer: display Ordered cylinders while in Goto mode
When flying an ordered task, the pilot may select to Goto a
turnpoint, but at the same time he wants to know where he is
in relation to the task.  For example, if he wants to nick the
turnpoint OZ and then land at the Goto point.
If the pilot does not want to see the ordered OZs, he can clear
the task from the task manager.