dlgConfiguration: remove Task Default Types screen
And save the task type defaults each time a task is committed in the
Task Manager instead.

By saving the task defaults every time we commit a task from the task
manager, we can remove these properties from being needed in the
configuration screens.
Also, move the AATMargin property to the from the Task Default Types
config screen to the Task Rules config screen because it does not exist in the
Task Manager.  Rename the remaining Task Config panel to "Contest."