dlgMatItemClick.cpp: new dialog when Mat TP is clicked
If flying a MAT task and a TP is clicked, then this dialog
appears instead of the normal MapListItem dialog.
It shows proforma task information if the TP is inserted
next, and allows the user to Add, More options, or Cancel.
If TP is already the next TP in the task, it prompts user
if he wants to remove from task.

Conflicts:
	src/Dialogs/Dialogs.h
	src/Dialogs/MapItemListDialog.cpp
	src/Task/MapTaskManager.cpp