Contest/Settings: new option to disable OLC optimisation
This a patch allows the user to disable OLC optimisation if they don't
require it.

While I was doing performance testing on the KOBO I discovered that
90% of the CPU time was spent running OLC calculations resulting in
ever other frame being greatly delayed on the draw.

For longer flights it could take several seconds to refresh the
screen, meaning that most of the info boxes were showing very delayed
information and the glider position was lagging.

This patch is quite simple and adds an option in the Task > Contest
Type box to allow a user to chose 'None'.  This will prevent any
calculation engine running and drops the CPU on the KOBO to less than
9% (and the CPU frequency down to minimum for most of the time) It
also ensures that the display continues to refresh nicely Patch should
apply to head.