AbstractTaskFactory: don't change finish size when mutating from a tp
If the last tp is added to the task, and the task is saved, it has
always changed the last tp to a FINISH type of point.  However, it should
not change the size of the TP because that might surprise the pilot
if he set the finish size already.